migrate main class
